gnhf 7: Extracted a shared HMAC signed-token primitive (lib/signed-token.ts), collapsing the byte-identical crypto and sign/verify skeletons duplicated across pull/offer.ts and quiz/quiz-offer.ts into thin wrappers.
